use ReadOnlyProperty instead of custom class

+++ http4k-cloudnative/src/test/kotlin/org/http4k/cloudnative/env/EnvironmentKeyTest.kt
@@ -10,9 +10,11 @@ import org.http4k.cloudnative.env.EnvironmentKey.k8s.serviceUriFor
 import org.http4k.core.Uri
 import org.http4k.core.with
 import org.http4k.lens.LensFailure
+import org.http4k.lens.Query
 import org.http4k.lens.composite
 import org.http4k.lens.int
 import org.http4k.lens.by
+import org.http4k.lens.long
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import java.util.Properties
@@ -78,7 +80,7 @@ class EnvironmentKeyTest {
 
     @Test
     fun `using property method`() {
-        val MY_GREAT_ENV_VARIABLE by EnvironmentKey.int().by().required()
+        val MY_GREAT_ENV_VARIABLE by EnvironmentKey.long().by().required()
         assertThat(MY_GREAT_ENV_VARIABLE(from("MY_GREAT_ENV_VARIABLE" to "123")), equalTo(123))
     }
